4.15 Move state in offers list to hoc
